Refactor Code: Add utils
